+++ Changelog.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Vivado variant autodetection to `ipb-prog vivado`.
 - `srcs run` command, to simplify running commands in the source area.
 - `VivadoConsole` can quite on Vivado's CRITICAL WARNINGS
+- `--tag` option to `vivado package` to customise the tarball name.
 
 ## [0.2.9] - 2018-01-08
 ### Changed